requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
requests==2.32.0
Pillow==10.3.0
together==0.2.5 